Async loop method replacement (from: run_forever( ) to: run_until_complete( ))
**UPDATES!**

the last commit was used the `run_forever()` method, however, if I added the command as a task in the `run_forever()` method, the function would always run, as if I put an entire onion in an activated blender with no lid. Now, when I add the function, it is executed and then exits the loop.

the help was provided by Octo Code Server (It's a brazilian server)